Add fixes to pass all Themis tests

+++ lab1/problem4.c
@@ -5,14 +5,14 @@
 int main() {
 
     int fs, n;
-    float A, f, phi;
-    scanf("%i %i %f %f %f", &fs, &n, &A, &f, &phi);
+    double A, f, phi;
+    scanf("%i %i %lf %lf %lf", &fs, &n, &A, &f, &phi);
 
     printf("[");
 
     for (int i = 0; i < n; ++i) {
-        float t = (float) i * (1 / (float) fs);
-        float result = A * cos(2 * pi * f * t + phi);
+        double t = (double) i * (1 / (double) fs);
+        double result = A * cos(2 * pi * f * t + phi);
         printf("%i", (int) result);
         if (i < n - 1)
             printf(",");
